style: changed training_funs into training_tools
- Change training_funs to tf_/torch_training_tools.py
- Move training tools to utils folder
- Rename train_keras_app to train_tf_model to align with PyTorch convention
  naming
