+++ tests/tests_tensorflow/utils/test_tf_training_tools.py
@@ -20,76 +20,77 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from torch.utils.data import DataLoader
-
-from oodeel.models.training_funs_torch import train_torch_model
-from tests.tests_torch import generate_data_torch
+from oodeel.utils.tf_training_tools import train_tf_model
+from tests.tests_tensorflow import generate_data_tf
 
 
 def test_convnet_classifier():
-    input_shape = (3, 32, 32)
+    input_shape = (32, 32, 3)
     num_labels = 10
     samples = 100
 
     train_config = {
         "model_name": "toy_convnet",
+        "batch_size": 5,
         "epochs": 3,
+        "input_shape": input_shape,
         "num_classes": num_labels,
-        "cuda_idx": None,
+        "is_prepared": False,
     }
 
-    data = generate_data_torch(
+    data = generate_data_tf(
         x_shape=input_shape, num_labels=num_labels, samples=samples, one_hot=False
     )
-    data = DataLoader(data, batch_size=samples // 2)
-    train_torch_model(data, **train_config)
+
+    train_tf_model(data, **train_config)
 
 
-def test_train_torch_model_imagenet():
-    input_shape = (3, 224, 224)
+def test_train_tf_model_imagenet():
+    input_shape = (224, 224, 3)
     num_labels = 1000
     samples = 100
 
     train_config = {
+        "batch_size": 5,
         "epochs": 3,
+        "input_shape": input_shape,
         "num_classes": num_labels,
-        "cuda_idx": None,
+        "is_prepared": False,
     }
 
-    data = generate_data_torch(
+    data = generate_data_tf(
         x_shape=input_shape, num_labels=num_labels, samples=samples, one_hot=False
     )
 
-    data = DataLoader(data, batch_size=samples // 2)
-    train_torch_model(
-        data, model_name="mobilenet_v2", imagenet_pretrained=True, **train_config
+    train_tf_model(
+        data, model_name="MobileNet", imagenet_pretrained=True, **train_config
     )
 
 
-def test_train_torch_model():
-    input_shape = (3, 56, 56)
+def test_train_tf_model():
+    input_shape = (56, 56, 3)
     num_labels = 123
     samples = 100
 
     train_config = {
+        "batch_size": 5,
         "epochs": 3,
+        "input_shape": input_shape,
         "num_classes": num_labels,
-        "cuda_idx": None,
+        "is_prepared": False,
     }
 
-    data = generate_data_torch(
+    data = generate_data_tf(
         x_shape=input_shape, num_labels=num_labels, samples=samples, one_hot=False
     )
 
-    validation_data = generate_data_torch(
+    validation_data = generate_data_tf(
         x_shape=input_shape, num_labels=num_labels, samples=samples, one_hot=False
     )
 
-    data = DataLoader(data, batch_size=samples // 2)
-    validation_data = DataLoader(validation_data, batch_size=samples // 2)
-    train_torch_model(
+    train_tf_model(
         data,
-        model_name="mobilenet_v2",
+        model_name="MobileNet",
         imagenet_pretrained=False,
         validation_data=validation_data,
         **train_config
